+++ src/CMakeLists.txt
@@ -54,6 +54,7 @@ file(GLOB_RECURSE SOURCES
     libsnark_helpers/libsnark_helpers.cpp
     types/bits.cpp
     util.cpp
+    pghr13_response.cpp
 )
 
 # Enable Boost for program_options
