+++ src/prover_server.cc
@@ -25,6 +25,8 @@
 // Include the file generated by gRPC
 #include "prover.grpc.pb.h"
 
+#include "pghr13_response.hpp"
+
 using grpc::Server;
 using grpc::ServerBuilder;
 using grpc::ServerContext;
@@ -38,191 +40,12 @@ using proverpkg::Prover;
 using proverpkg::EmptyMessage;
 using proverpkg::PackedDigest;
 using proverpkg::ProofInputs;
-using proverpkg::HexadecimalPointBaseGroup1Affine;
-using proverpkg::HexadecimalPointBaseGroup2Affine;
-using proverpkg::VerificationKey;
-using proverpkg::ExtendedProof;
-using proverpkg::R1csPpzksnarkExtendedProof;
-using proverpkg::R1csPpzksnarkVerificationKey;
 
 typedef libff::default_ec_pp ppT;
 typedef libff::Fr<ppT> FieldT;
 typedef sha256_ethereum<FieldT> HashT;
 
-libsnark::merkle_authentication_node ParseMerkleNode(std::string mk_node) {
-  return libff::bit_vector(libzeth::hexadecimal_digest_to_binary_vector(mk_node));
-}
-
-libzeth::ZethNote ParseZethNote(const proverpkg::ZethNote& note) {
-  bits256 noteAPK = libzeth::hexadecimal_digest_to_bits256(note.apk());
-  bits64 noteValue = libzeth::hexadecimal_value_to_bits64(note.value());
-  bits256 noteRho = libzeth::hexadecimal_digest_to_bits256(note.rho());
-  bits384 noteTrapR = libzeth::get_bits384_from_vector(libzeth::hexadecimal_str_to_binary_vector(note.trapr()));
-
-  return libzeth::ZethNote(
-    noteAPK,
-    noteValue,
-    noteRho,
-    noteTrapR
-  );
-}
-
-libzeth::JSInput ParseJSInput(const proverpkg::JSInput& input) {
-  if (ZETH_MERKLE_TREE_DEPTH != input.merklenode_size()) {
-    throw std::invalid_argument("Invalid merkle path length");
-  }
-
-  libzeth::ZethNote inputNote = ParseZethNote(input.note());
-  size_t inputAddress = input.address();
-  bitsAddr inputAddressBits = libzeth::get_bitsAddr_from_vector(libzeth::address_bits_from_address(inputAddress, ZETH_MERKLE_TREE_DEPTH));
-  bits256 inputSpendingASK = libzeth::hexadecimal_digest_to_bits256(input.spendingask());
-  bits256 inputNullifier = libzeth::hexadecimal_digest_to_bits256(input.nullifier());
-
-  std::vector<libsnark::merkle_authentication_node> inputMerklePath;
-  for(int i = 0; i < ZETH_MERKLE_TREE_DEPTH; i++) {
-    libsnark::merkle_authentication_node mk_node = ParseMerkleNode(input.merklenode(i));
-    inputMerklePath.push_back(mk_node);
-  }
-
-  return libzeth::JSInput(
-    inputMerklePath,
-    inputAddress,
-    inputAddressBits,
-    inputNote,
-    inputSpendingASK,
-    inputNullifier
-  );
-}
-
-HexadecimalPointBaseGroup1Affine FormatHexadecimalPointBaseGroup1Affine(libff::alt_bn128_G1 point) {
-  libff::alt_bn128_G1 aff = point;
-  aff.to_affine_coordinates();
-  std::string xCoord = "0x" + HexStringFromLibsnarkBigint(aff.X.as_bigint());
-  std::string yCoord = "0x" + HexStringFromLibsnarkBigint(aff.Y.as_bigint());
-
-  HexadecimalPointBaseGroup1Affine res;
-  res.set_xcoord(xCoord);
-  res.set_ycoord(yCoord);
-
-  return res;
-}
-
-HexadecimalPointBaseGroup2Affine FormatHexadecimalPointBaseGroup2Affine(libff::alt_bn128_G2 point) {
-  libff::alt_bn128_G2 aff = point;
-  aff.to_affine_coordinates();
-  std::string xC1Coord = "0x" + HexStringFromLibsnarkBigint(aff.X.c1.as_bigint());
-  std::string xC0Coord = "0x" + HexStringFromLibsnarkBigint(aff.X.c0.as_bigint());
-  std::string yC1Coord = "0x" + HexStringFromLibsnarkBigint(aff.Y.c1.as_bigint());
-  std::string yC0Coord = "0x" + HexStringFromLibsnarkBigint(aff.Y.c0.as_bigint());
-
-  HexadecimalPointBaseGroup2Affine res;
-  res.set_xc0coord(xC0Coord);
-  res.set_xc1coord(xC1Coord);
-  res.set_yc0coord(yC0Coord);
-  res.set_yc1coord(yC1Coord);
-
-  return res;
-}
-
-void PrepareProofResponse(extended_proof<ppT>& ext_proof, ExtendedProof* message) {
-  libsnark::r1cs_ppzksnark_proof<ppT> proofObj = ext_proof.get_proof();
-
-  HexadecimalPointBaseGroup1Affine *a = new HexadecimalPointBaseGroup1Affine();
-  HexadecimalPointBaseGroup1Affine *a_p = new HexadecimalPointBaseGroup1Affine();
-  HexadecimalPointBaseGroup2Affine *b = new HexadecimalPointBaseGroup2Affine(); // in G2
-  HexadecimalPointBaseGroup1Affine *b_p = new HexadecimalPointBaseGroup1Affine();
-  HexadecimalPointBaseGroup1Affine *c = new HexadecimalPointBaseGroup1Affine();
-  HexadecimalPointBaseGroup1Affine *c_p = new HexadecimalPointBaseGroup1Affine();
-  HexadecimalPointBaseGroup1Affine *h = new HexadecimalPointBaseGroup1Affine();
-  HexadecimalPointBaseGroup1Affine *k = new HexadecimalPointBaseGroup1Affine();
-
-  a->CopyFrom(FormatHexadecimalPointBaseGroup1Affine(proofObj.g_A.g));
-  a_p->CopyFrom(FormatHexadecimalPointBaseGroup1Affine(proofObj.g_A.h));
-  b->CopyFrom(FormatHexadecimalPointBaseGroup2Affine(proofObj.g_B.g)); // in G2
-  b_p->CopyFrom(FormatHexadecimalPointBaseGroup1Affine(proofObj.g_B.h));
-  c->CopyFrom(FormatHexadecimalPointBaseGroup1Affine(proofObj.g_C.g));
-  c_p->CopyFrom(FormatHexadecimalPointBaseGroup1Affine(proofObj.g_C.h));
-  h->CopyFrom(FormatHexadecimalPointBaseGroup1Affine(proofObj.g_H));
-  k->CopyFrom(FormatHexadecimalPointBaseGroup1Affine(proofObj.g_K));
-
-  libsnark::r1cs_ppzksnark_primary_input<ppT> pubInputs = ext_proof.get_primary_input();
-  std::stringstream ss;
-  ss << "[";
-  for (size_t i = 0; i < pubInputs.size(); ++i) {
-    ss << "\"0x" << HexStringFromLibsnarkBigint(pubInputs[i].as_bigint()) << "\"";
-    if ( i < pubInputs.size() - 1 ) {
-        ss<< ", ";
-    }
-  }
-  ss << "]";
-  std::string inputs_json = ss.str();
-
-  // Note on memory safety: set_allocated deleted the allocated objects
-  // See: https://stackoverflow.com/questions/33960999/protobuf-will-set-allocated-delete-the-allocated-object
-  R1csPpzksnarkExtendedProof *r1csPpzksnarkExtendedProof = message->mutable_r1csppzksnarkextendedproof();
-  
-  r1csPpzksnarkExtendedProof->set_allocated_a(a);
-  r1csPpzksnarkExtendedProof->set_allocated_ap(a_p);
-  r1csPpzksnarkExtendedProof->set_allocated_b(b);
-  r1csPpzksnarkExtendedProof->set_allocated_bp(b_p);
-  r1csPpzksnarkExtendedProof->set_allocated_c(c);
-  r1csPpzksnarkExtendedProof->set_allocated_cp(c_p);
-  r1csPpzksnarkExtendedProof->set_allocated_h(h);
-  r1csPpzksnarkExtendedProof->set_allocated_k(k);
-  r1csPpzksnarkExtendedProof->set_inputs(inputs_json);
-}
-
-void PrepareVerifyingKeyResponse(libsnark::r1cs_ppzksnark_verification_key<ppT>& vk, VerificationKey* message) {
-  HexadecimalPointBaseGroup2Affine *a = new HexadecimalPointBaseGroup2Affine(); // in G2
-  HexadecimalPointBaseGroup1Affine *b = new HexadecimalPointBaseGroup1Affine(); // in G1
-  HexadecimalPointBaseGroup2Affine *c = new HexadecimalPointBaseGroup2Affine(); // in G2
-  HexadecimalPointBaseGroup2Affine *g = new HexadecimalPointBaseGroup2Affine(); // in G2
-  HexadecimalPointBaseGroup1Affine *gb1 = new HexadecimalPointBaseGroup1Affine(); // in G1
-  HexadecimalPointBaseGroup2Affine *gb2 = new HexadecimalPointBaseGroup2Affine(); // in G2
-  HexadecimalPointBaseGroup2Affine *z = new HexadecimalPointBaseGroup2Affine(); // in G2
-
-  a->CopyFrom(FormatHexadecimalPointBaseGroup2Affine(vk.alphaA_g2)); // in G2
-  b->CopyFrom(FormatHexadecimalPointBaseGroup1Affine(vk.alphaB_g1)); // in G1
-  c->CopyFrom(FormatHexadecimalPointBaseGroup2Affine(vk.alphaC_g2)); // in G2
-  g->CopyFrom(FormatHexadecimalPointBaseGroup2Affine(vk.gamma_g2)); // in G2
-  gb1->CopyFrom(FormatHexadecimalPointBaseGroup1Affine(vk.gamma_beta_g1)); // in G1
-  gb2->CopyFrom(FormatHexadecimalPointBaseGroup2Affine(vk.gamma_beta_g2)); // in G2
-  z->CopyFrom(FormatHexadecimalPointBaseGroup2Affine(vk.rC_Z_g2)); // in G2
-
-  std::stringstream ss;
-  unsigned icLength = vk.encoded_IC_query.rest.indices.size() + 1;
-  ss <<  "[[" << outputPointG1AffineAsHex(vk.encoded_IC_query.first) << "]";
-  for (size_t i = 1; i < icLength; ++i) {
-    auto vkICi = outputPointG1AffineAsHex(vk.encoded_IC_query.rest.values[i - 1]);
-    ss << ",[" <<  vkICi << "]";
-  }
-  ss << "]";
-  std::string IC_json = ss.str();
-
-  // Note on memory safety: set_allocated deleted the allocated objects
-  // See: https://stackoverflow.com/questions/33960999/protobuf-will-set-allocated-delete-the-allocated-object
-  R1csPpzksnarkVerificationKey *r1csPpzksnarkVerificationKey = message->mutable_r1csppzksnarkverificationkey();
-
-  r1csPpzksnarkVerificationKey->set_allocated_a(a);
-  r1csPpzksnarkVerificationKey->set_allocated_b(b);
-  r1csPpzksnarkVerificationKey->set_allocated_c(c);
-  r1csPpzksnarkVerificationKey->set_allocated_g(g);
-  r1csPpzksnarkVerificationKey->set_allocated_gb1(gb1);
-  r1csPpzksnarkVerificationKey->set_allocated_gb2(gb2);
-  r1csPpzksnarkVerificationKey->set_allocated_z(z);
-  r1csPpzksnarkVerificationKey->set_ic(IC_json);
-}
-
-/*
-void PrepareProofResponse(extended_proof<ppT, libsnark::r1cs_gg_ppzksnark_proof<ppT>>& ext_proof, ExtendedProof* message){
-  
-}
-
-void PrepareVerifyingKeyResponse(libsnark::r1cs_gg_ppzksnark_verification_key<ppT>& vk, VerificationKey* message) {
-
-}
-*/
-class ProverImpl final : public Prover::Service {//TODO abstract this
+class ProverImpl final : public Prover::Service {
 private:
   libzeth::CircuitWrapper<ZETH_NUM_JS_INPUTS, ZETH_NUM_JS_OUTPUTS> prover;
   keyPairT<ppT> keypair; // Result of the setup
