+++ src/util.hpp
@@ -5,7 +5,13 @@
 #include <string>
 #include <cstdint>
 
+#include <libsnark/common/data_structures/merkle_tree.hpp>
+
 #include "types/bits.hpp"
+#include "types/note.hpp"
+#include "types/joinsplit.hpp"
+
+#include "prover.grpc.pb.h"
 
 namespace libzeth {
 
@@ -19,6 +25,11 @@ bits64 hexadecimal_value_to_bits64(std::string value_hex_str);
 
 std::vector<bool> convert_int_to_binary(int x);
 std::vector<bool> address_bits_from_address(int address, int tree_depth);
+
+//message parsing utils
+libsnark::merkle_authentication_node ParseMerkleNode(std::string mk_node);
+libzeth::ZethNote ParseZethNote(const proverpkg::ZethNote& note);
+libzeth::JSInput ParseJSInput(const proverpkg::JSInput& input);
 
 } // libzeth
